Create program with basic gui
Created a Python script with a gui that scrapes anilist.co for the top 50 trending anime shows in the year and season specified by the user. After the script runs, the gui displays a list of the top 50 shows and suggests one at random.

Currently on a basic tkinter gui, but will be upgraded shortly to a custom gui. Also will be adding features like re-roll, delete and have it display the genre and rating of the show.