Create CustomTkinter GUI
Created a custom gui for the program in customtkinter with a touch of vanilla tkinter.

Added a try-except block to the script which processes API response data and prepares to handle unexpected data formats. (implemented after NoneType error)

Added extra functionality with an option to delete already watched shows by selecting a show from the list and pressing button.

Added a reroll button which allows the user to receive a new show suggestion.

Installed and froze customtkinter library to requirements.txt.

Updated old gui.py to simplify data processed for Listbox element to list of strings instead of a list of dictionaries with detailed metadata.